optweight: Targeted Stable Balancing Weights Using Optimization

Use optimization to estimate weights that balance covariates for binary, multi-category, continuous, and multivariate treatments in the spirit of Zubizarreta (2015) doi:10.1080/01621459.2015.1023805 . The degree of balance can be specified for each covariate. In addition, sampling weights can be estimated that allow a sample to generalize to a population specified with given target moments of covariates.

See also

Author

Maintainer: Noah Greifer noah.greifer@gmail.com (ORCID)